LayoutTests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-url-02.html

<!DOCTYPE html>
<html>
<head>
<script src='../resources/multiple-iframe-plugin-test.js'></script>
<script>
var tests = [
    [false, 'plugin-types text/plain'],
    [false, 'plugin-types application/x-shockwave-flash', '/plugins/resources/mock-plugin.pl?url-doesnt-match-csp', 'application/x-webkit-test-netscape'],
    [false, 'plugin-types application/x-unknown-type', '/plugins/resources/mock-plugin-unknown-type.pl?type-attribute-doesnt-match-csp', 'application/x-webkit-test-netscape'],
];
</script>
</head>
<body onload="test()">
    <p>
        This tests our handling of non-`data:` URLs, given a `plugin-types` CSP
        directive. Consider this test passing if none of the following frames
        contains "FAIL" and four sets of console logs appear above.
    </p>
